git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/2011-12-15-fault@4575 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisst3DUserInterface/ui3VisibleList.h
@@ -71,7 +71,7 @@ public:
         return this->Objects.back();
     }
     
-    unsigned int size(void) const  {
+    size_t size(void) const  {
         return this->Objects.size();
     }
